# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Recipe Book",
    "description": "API System for Recipe Book",
    "version": "1.0.0",
    "contact": {
      "name": "CSE341-GROUP08 Project Github Repo Link",
      "url": "https://github.com/chrmux/cse341-group08-project"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/keywords/{id}": {
      "get": {
        "tags": [
          "Keywords"
        ],
        "summary": "Fetch keyword by ID",
        "description": "Find one from the database",
        "operationId": "getKeywordById",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": "true",
          "description": "id of the keyword",
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "put": {
        "tags": [
          "Keywords"
        ],
        "summary": " Update Keyword By ID",
        "description": "Update the existing keyword in database",
        "operationId": "updateKeywords",
        "parameters": [{
            "name": "id",
            "in": "path",
            "required": "true",
            "description": "id of the keyword",
            "type": "string"
          },
          {
            "name": "keyword",
            "description": "Keyword params to update the keyword",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "classification": {
                  "type": "string",
                  "example": "any"
                },
                "name": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Keyword has been Updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Keywords"
        ],
        "summary": " Delete keyword By ID",
        "description": "Delete keyword from the database",
        "operationId": "deleteKeywords",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": "true",
          "description": "id of the keyword",
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/ingredients/{id}": {
      "get": {
        "tags": [
          "Ingredients"
        ],
        "summary": "Fetch Ingredient",
        "description": "Find one from the database",
        "operationId": "getIngredientsById",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": "true",
          "description": "id of the Ingredient",
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "put": {
        "tags": [
          "Ingredients"
        ],
        "summary": "Update Ingredients",
        "description": "Update the existing ingredient in database",
        "operationId": "updateIngredient",
        "parameters": [{
            "name": "id",
            "in": "path",
            "required": "true",
            "description": "id of the Ingredient",
            "type": "string"
          },
          {
            "name": "ingredient",
            "description": "Ingredient params to update the ingredient",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string",
                  "example": "any"
                },
                "data": {
                  "type": "string",
                  "example": "any"
                },
                "category": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "Ingredients"
        ],
        "summary": "Delete Ingredient",
        "description": "Delete ingredient from the database",
        "operationId": "deleteIngredient",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "description": "id of the Ingredient",
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/photos/{id}": {
      "get": {
        "tags": [
          "Photo"
        ],
        "summary": "Fetch Photo By ID",
        "description": "Find one from the database",
        "operationId": "getPhotoById",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": "true",
          "description": "id of the Photo",
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "put": {
        "tags": [
          "Photo"
        ],
        "summary": "Update Photo By ID",
        "description": "Update the existing Photo in database",
        "operationId": "updatePhoto",
        "parameters": [{
            "name": "id",
            "in": "path",
            "required": "true",
            "description": "id of the Photo",
            "type": "string"
          },
          {
            "name": "photo",
            "description": "Photo params to update the photo",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "title": {
                  "type": "string",
                  "example": "any"
                },
                "photoUrl": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "Photo"
        ],
        "summary": "Delete Photo By ID",
        "description": "Delete photo from the database",
        "operationId": "deletePhoto",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": "true",
          "description": "id of the Photo",
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
...
}